aiuto esplode tutto

+++ Milestone_3/Minishell/libft/ft_atoi.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yoherfan <yoherfan@student.42.fr>          +#+  +:+       +#+        */
+/*   By: cciapett <cciapett@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/19 18:04:33 by yoherfan          #+#    #+#             */
-/*   Updated: 2024/12/04 16:40:40 by yoherfan         ###   ########.fr       */
+/*   Updated: 2025/06/17 14:45:30 by cciapett         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+long long int	ft_atoi(const char *nptr)
 {
-	int	i;
-	int	result;
-	int	sign;
+	int				i;
+	long long int	result;
+	int				sign;
 
 	sign = 0;
 	result = 0;
